requirements.txt
python-dotenv
py-cord[speed]==2.0.0rc1
pylast
requests